% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divorce_margarine.R
\docType{data}
\name{divorce_margarine}
\alias{divorce_margarine}
\title{Divorce rate and margarine consumption data}
\format{An object of class \code{"data.frame"}.}
\source{
\url{http://www.tylervigen.com/spurious-correlations}{Spurious Correlations}
data(divorce_margarine)
with(divorce_margarine, plot(argarine_consumption_per_capita, divorce_rate_maine))
}
\usage{
data(divorce_margarine)
}
\description{
Divorce rates in Maine and per capita consumption of margarine in US data
}
\details{
\itemize{
divorce_rate_maine. ivorce rates in Maine
margarine_consumption_per_capita. per capita consumption of margarine in US.
year. Year.
}
}
\keyword{datasets}
