% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_theme_set.R
\name{ds_theme_set}
\alias{ds_theme_set}
\title{dslabs theme set}
\usage{
ds_theme_set(new = "theme_minimal", args = NULL, base_size = 11,
  base_family = "Helvetica", bold_title = TRUE, ...)
}
\arguments{
\item{new}{a prebuilt theme. Defaults to theme_minimal}

\item{args}{the arguments to be passed along to the theme function. Defaults to \code{"NULL"}.}

\item{base_size}{if \code{"args"} is \code{"NULL"}, base_size is one of the arguments passed to the theme function. It defaults to 11.}

\item{base_family}{if \code{"args"} is \code{"NULL"}, base_family is one of the arguments passed to the theme function. It defaults to Helvetica.}

\item{bold_title}{it TRUE, sets titles to be bold}

\item{...}{additional arguments to be used by theme}
}
\value{
None
}
\description{
Called without arguments, this function sets a ggplot2 theme used throughout the data science labs.
}
\examples{
library(ggplot2)
ds_theme_set()
qplot(hp, mpg, data=mtcars, color=am, facets=gear~cyl, 
main="Scatterplots of MPG vs. Horsepower", 
xlab="Horsepower", ylab="Miles per Gallon")

}

