% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapminder.R
\docType{data}
\name{gapminder}
\alias{gapminder}
\alias{oecd}
\alias{opec}
\title{Gapminder Data}
\format{An object of class \code{"data.frame"}.}
\usage{
data(gapminder)
}
\description{
Health and income outcomes for 184 countries from 1960 to 2016. Also includes two character verctors, oecd and opec, of OECD and OPEC countries from 2016.
}
\details{
\itemize{
 \item country.  
 \item year.
 \item infant_mortality. Infant deaths per 1000.
 \item life_expectancy. Life expectancy in years.
 \item fertility. Average number of children per woman. 
 \item population. Country population.
 \item income. Average income for the country.
 \item gpd_per_capita. GDP per capita for the country.
 \item region. Region of the world.
}
}
\examples{
data(gapminder)
head(gapminder)
print(oecd)
print(opec)
}
\keyword{datasets}

