% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heights.R
\docType{data}
\name{heights}
\alias{heights}
\title{Self-reported Heights in Inches}
\format{
An object of class \code{"data.frame"}.
}
\usage{
heights
}
\description{
Self-reported heights and sex. The heights were converted to inches from the original data included in \code{\link{reported_heights}}.
}
\details{
\itemize{
 \item sex. A factor with the self-reported sex. 
 \item height. A numeric vector with self-reported heights in inches. 
}
}
\examples{
head(heights)
}
\seealso{
\code{\link{reported_heights}} for the original data source.
}
\keyword{datasets}
