% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice_weights.R
\docType{data}
\name{mice_weights}
\alias{mice_weights}
\title{Mice weights}
\format{
An object of class \code{"data.frame"}.
}
\source{
Karen Svenson from Jackson Laboratories
}
\usage{
mice_weights
}
\description{
Weights for mice under two diets: chow and high fat. Data provided by Karen Svenson from Jackson Laboratories. Funding to generate these data came from NIH grant P50 GM070683 awarded to Gary Churchill.
}
\details{
\itemize{
  \item sex. The sex of the mice.
  \item diet. The diet of the mice: chow or high fat.
  \item body_weight. Body weight in grams at 12 weeks.
}
}
\examples{
mice_weights |> head()
with(mice_weights, table(sex, diet))

}
\keyword{datasets}
