% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movielens.R
\docType{data}
\name{movielens}
\alias{movielens}
\title{Movie ratings}
\format{Two object of class \code{data.frame}.}
\source{
\href{http://files.grouplens.org/datasets/movielens/ml-latest-small.zip}{http://files.grouplens.org/datasets/movielens/ml-latest-small.zip}
}
\usage{
data(movielens)
}
\description{
MovieLens Latest Dataset (Small)
}
\details{
\itemize{
\item movieId. Unique ID for the movie.
\item title. Movie title (not unique). 
\item year. Year the movie was released.
\item genres. Genres associated with the movies.
\item userId. Unique ID for the user. 
\item rating. A rating between 0 and 5 for the movie.
\item timestap. Date and time the rating was given.
}
}
\examples{
data(movielens)
head(movielens)
}
\keyword{(December}
\keyword{(TiiS)}
\keyword{19}
\keyword{2015),}
\keyword{2015.}
\keyword{4,}
\keyword{5,}
\keyword{@references}
\keyword{A.}
\keyword{ACM}
\keyword{Article}
\keyword{Context.}
\keyword{DOI=http://dx.doi.org/10.1145/2827872}
\keyword{Datasets:}
\keyword{F.}
\keyword{Harper}
\keyword{History}
\keyword{Intelligent}
\keyword{Interactive}
\keyword{Joseph}
\keyword{Konstan.}
\keyword{Maxwell}
\keyword{MovieLens}
\keyword{Systems}
\keyword{The}
\keyword{Transactions}
\keyword{and}
\keyword{datasets}
\keyword{on}
\keyword{pages.}
