% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/murders.R
\docType{data}
\name{murders}
\alias{murders}
\title{US gun murders by state for 2010}
\format{An object of class \code{"data.frame"}.}
\source{
\href{https://en.wikipedia.org/wiki/Murder_in_the_United_States_by_state}{Wikipedia}
}
\usage{
data(murders)
}
\description{
Gun murder data are from FBI reports. Also contains the population of each state.
}
\details{
\itemize{
 \item state. US state
 \item abb. Abbreviation of US state
 \item region. Geographical US region
 \item population. State population (2010)
 \item total. Number of gun murders in state (2010)
}
}
\examples{
data(murders)
print(murders)
}
\keyword{datasets}
