% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olives.R
\docType{data}
\name{olives}
\alias{olives}
\title{Italian olive}
\format{An object of class \code{"data.frame"}.}
\source{
J. Zupan, and J. Gasteiger.  Neural Networks in Chemistry and Drug Design.
}
\usage{
data(olives)
}
\description{
Percentage composition of 8 fatty acids found in the lipid fraction of 572 Italian olive oils
}
\details{
\itemize{
\item area. Area of Italy. 
\item region. General region of Italy 
\item palmitic. Percent of palmitic acidin sample. 
\item palmitoleic. Percent of palmitoleic in sample. 
\item stearic. Percent of stearic acid in sample.  
\item oleic. Percent of oleic acid in sample.  
\item linoleic. Percent of linoleic acid in sample.  
\item linolenic. Percent of linolenic acid in sample.  
\item arachidic. Percent of arachidic acid in sample.  
\item eicosenoic. Percent of eicosenoic acid in sample.  
}
}
\examples{
data(olives)
head(olives)
}
\keyword{datasets}
