% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_example.R
\docType{data}
\name{outlier_example}
\alias{outlier_example}
\title{Adulte male heights in feet with outlier}
\format{An object of class \code{"numeric"}.}
\usage{
data(outlier_example)
}
\description{
This dataset was randomly generated with a normal distribution with average 5 feet nine inches and standard deviation 3 inches. One value was changed to be mistakenly reported in centimeters rather than feet.
}
\examples{
data(outlier_example)
mean(outlier_example)
median(outlier_example)
}
\keyword{datasets}

