% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polls_us_election_2016.R
\docType{data}
\name{polls_us_election_2016}
\alias{polls_us_election_2016}
\title{Fivethirtyeight 2016 Poll Data}
\format{An object of class \code{"data.frame"}.}
\source{
\href{https://projects.fivethirtyeight.com/2016-election-forecast/?ex_cid=rrpromo}{Fivethirtyeight}
}
\usage{
data(polls_us_election_2016)
}
\description{
Poll results from US 2016 presidential elections aggregated from HuffPost Pollster, RealClearPolitics, polling firms and news reports. Original csv file is here: http://projects.fivethirtyeight.com/general-model/president_general_polls_2016.csv
}
\details{
\itemize{
\item state. State in which poll was taken. `U.S` is for national polls.
\item startdate. Poll's start data.
\item enddate. Poll's end data.
\item pollster. Pollster conducting the poll
\item grade. Grade assigned by fivethirtyeight to pollster.
\item samplesize. Sample size. 
\item population. Type of population being polled.
\item rawpoll_clinton. Percentage for Hillary Clinton.
\item rawpoll_trump. Percentage for Donald Trump
\item rawpoll_johnson. Percentage for Gary Johnson
\item rawpoll_mcmullin. Percentage for Evan McCullin.
}
}
\examples{
data(polls_us_election_2016)
head(polls_us_election_2016)
}
\keyword{datasets}

