% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_death_counts.R
\docType{data}
\name{pr_death_counts}
\alias{pr_death_counts}
\title{Puerto Rico daily mortality}
\format{
An object of class \code{data.frame}.
}
\source{
Puerto Rico Demographic Registry. Data was extracted from PDF provided in `system.file("extdata", "RD-Mortality-Report_2015-18-180531.pdf", package = "dslabs")`
}
\usage{
pr_death_counts
}
\description{
A data frame with Puerto Rico daily mortality counts 2015 to May 2018.
This includes the day hurricanes Maria made 2017-09-20.
}
\details{
\itemize{
\item date. Date of the count.
\item deaths. Number of deaths reported that day.
}
}
\examples{

with(pr_death_counts, plot(date, deaths))

}
\keyword{datasets}
