% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/research_funding_rates.R
\docType{data}
\name{research_funding_rates}
\alias{research_funding_rates}
\alias{raw_data_research_funding_rates}
\title{Gender bias in research funding in the Netherlands}
\format{
An object of class \code{"data.frame"}.
}
\source{
van der Lee and Ellemers (2015) PNAS \url{https://www.pnas.org/content/112/40/12349.abstract/}
}
\usage{
research_funding_rates
}
\description{
Table S1 from paper title "Gender contributes to personal research funding success in The Netherlands"
}
\details{
\itemize{
  \item discipline. Research area discipline.
  \item applications_total. Total applications.
  \item applications_men. Total applications by men.
  \item applications_women. Total applications by women.
  \item awards_total. Total awards.
  \item awards_men. Total awards received by men.
  \item awards_women. Total awards received by women.
  \item success_rates_total. Overall success rate.
  \item success_rates_men. Success rate for men.
  \item success_rates_women. Success rate for women.
}
}
\examples{
research_funding_rates
# The raw data for this table is available from
invisible(raw_data_research_funding_rates)

}
\keyword{datasets}
