% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_us_election_2012.R
\docType{data}
\name{results_us_election_2012}
\alias{results_us_election_2012}
\title{2012 US presidential election results}
\format{
An object of class \code{"data.frame"}.
}
\source{
Wikepedia: \url{https://en.wikipedia.org/w/index.php?title=2012_United_States_presidential_election&oldid=1264588444}
}
\usage{
results_us_election_2012
}
\description{
Percentages for the four major candidates by state in the US 2012 presidential elections. It includes congressional districts
for Maine and Nebraska.
}
\details{
\itemize{
\item state. State in which poll was taken. `U.S` is for national polls.
\item electoral_votes. Electoral votes for that state.
\item obama. Percent obtained by Barack Obama.
\item romney. Percent obtained by Mitt Romney.
\item johnson. Percent obtained by Gary Johnson.
\item stein. Percent obtained by Jill Stein.
}
}
\examples{
head(results_us_election_2012)
}
\keyword{datasets}
