% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/take_poll.R
\name{take_poll}
\alias{take_poll}
\title{Models results from taking a poll}
\usage{
take_poll(n, ...)
}
\arguments{
\item{n}{the sample size}

\item{...}{additional arguments to be used by the function that generates the}
}
\value{
None
}
\description{
The function shows a plot of a random sample from an urn of blue and red beads. The sample is taken with replacement. The proportion of blue beads is not shown so that students can try to estimate it.
}
\examples{
library(ggplot2)
take_poll(25)

}

