% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/us_contagious_disease.R
\docType{data}
\name{us_contagious_diseases}
\alias{us_contagious_diseases}
\title{Contagious Disease Data for US States}
\format{An object of class \code{"data.frame"}.}
\source{
\href{http://www.tycho.pitt.edu/}{Tycho Project}
}
\usage{
data(us_contagious_diseases)
}
\description{
Yearly counts for Hepatitis A Measles, Mumps, Pertussis, Polio,  Rubella, and Small for US states. Original data courtesy of Tycho Project (http://www.tycho.pitt.edu/).
}
\details{
#' \itemize{
 \item disease. A factor containing name of disease.
 \item state. A factor containing stata name.
 \item year. 
 \item weeks_reporting. Number of weeks that reported counts that year.
 \item count. Total number of reported cases. 
 \item population. State population that year interpolated for non census years.
}
}
\examples{
data(us_contagious_diseases)
head(us_contagious_diseases)
}
\keyword{datasets}

