% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admissions.R
\docType{data}
\name{admissions}
\alias{admissions}
\title{Gender bias among graduate school admissions to UC Berkeley.}
\format{
An object of class \code{"data.frame"}.
}
\source{
PJ Bickel, EA Hammel, and JW O'Connell. Science (1975)
}
\usage{
admissions
}
\description{
The admission data for six majors for the fall of 1973; often used as an example of Simpson's paradox
}
\details{
\itemize{
  \item major. The major or university department.
  \item gender. Men or women.
  \item admitted. Percent of students admitted.
  \item applicants. Total number of applicants.
}
}
\examples{
admissions
}
\keyword{datasets}
