% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divorce_margarine.R
\docType{data}
\name{divorce_margarine}
\alias{divorce_margarine}
\title{Divorce rate and margarine consumption data}
\format{
An object of class \code{"data.frame"}.
}
\source{
\href{https://www.tylervigen.com/spurious-correlations}{Spurious Correlations}
}
\usage{
divorce_margarine
}
\description{
Divorce rates in Maine and per capita consumption of margarine in US data
}
\details{
\itemize{
  \item divorce_rate_maine. Divorce per 1000 in Maine.
  \item margarine_consumption_per_capita. US per capita consumption of margarine in pounds.
  \item year. Year.
}
}
\examples{
with(divorce_margarine, plot(margarine_consumption_per_capita, divorce_rate_maine))
}
\keyword{datasets}
