% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_theme_set.R
\name{ds_theme_set}
\alias{ds_theme_set}
\title{dslabs theme set}
\usage{
ds_theme_set(
  new = "theme_bw",
  args = NULL,
  base_size = 11,
  bold_title = TRUE,
  ...
)
}
\arguments{
\item{new}{a prebuilt ggplot2 theme. Defaults to \code{"theme_minimal"}}

\item{args}{the arguments to be passed along to the ggplot2 theme function. Defaults to \code{"NULL"}.}

\item{base_size}{if \code{"args"} is \code{"NULL"}, base_size is one of the arguments passed to the theme function. It defaults to 11.}

\item{bold_title}{if TRUE, sets titles to be bold}

\item{...}{additional arguments to be used by theme}
}
\value{
None
}
\description{
This function sets a ggplot2 theme used throughout the data science labs. It can be called without arguments.
}
\examples{
library(ggplot2)
ds_theme_set()
qplot(hp, mpg, data=mtcars, color=am, facets=gear~cyl, 
main="Scatterplots of MPG vs. Horsepower", 
xlab="Horsepower", ylab="Miles per Gallon")

}
