% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heights.R
\docType{data}
\name{heights}
\alias{heights}
\title{Self-Reported Heights}
\format{
An object of class \code{"data.frame"}.
}
\usage{
heights
}
\description{
Self-reported heights in inches for males and females.
}
\details{
\itemize{
 \item sex. Male or Female.
 \item height. Height in inches. 
}
}
\examples{
head(heights)
}
\keyword{datasets}
