% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/murders.R
\docType{data}
\name{murders}
\alias{murders}
\title{US gun murder by state for 2010}
\format{An object of class \code{"data.frame"}.}
\source{
\href{https://en.wikipedia.org/wiki/Murder_in_the_United_States_by_state}{Wikipedia}
}
\usage{
data(murders)
}
\description{
Totals are from FBI reports. Population of each state for 2010.
}
\examples{
data(murders)
print(murders)
}
\keyword{datasets}

