% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_example.R
\docType{data}
\name{na_example}
\alias{na_example}
\title{Count data with some missing values}
\format{
An object of class \code{"integer"}.
}
\usage{
na_example
}
\description{
This dataset was randomly generated.
}
\examples{
print(sum(is.na(na_example)))
}
\keyword{datasets}
