% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olive.R
\docType{data}
\name{olive}
\alias{olive}
\title{Italian olive}
\format{
An object of class \code{"data.frame"}.
}
\source{
J. Zupan, and J. Gasteiger.  Neural Networks in Chemistry and Drug Design.
}
\usage{
olive
}
\description{
Composition in percentage of eight fatty acids found in the lipid fraction of 572 Italian olive oils
}
\details{
\itemize{
\item region. General region of Italy.
\item area. Area of Italy. 
\item palmitic. Percent palmitic acid of sample. 
\item palmitoleic. Percent palmitoleic of sample. 
\item stearic. Percent stearic acid of sample.  
\item oleic. Percent oleic acid of sample.  
\item linoleic. Percent linoleic acid of sample.  
\item linolenic. Percent linolenic acid of sample.  
\item arachidic. Percent arachidic acid of sample.  
\item eicosenoic. Percent eicosenoic acid of sample.  
}
}
\examples{
head(olive)
}
\keyword{datasets}
