% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_example.R
\docType{data}
\name{outlier_example}
\alias{outlier_example}
\title{Adult male heights in feet with outliers}
\format{
An object of class \code{"numeric"}.
}
\usage{
outlier_example
}
\description{
This dataset was randomly generated with a normal distribution (average: 5 feet 9 inches, standard deviation: 3 inches). One value was changed to be mistakenly reported in centimeters rather than feet.
}
\examples{
mean(outlier_example)
median(outlier_example)
}
\keyword{datasets}
