% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polls_2008.R
\docType{data}
\name{polls_2008}
\alias{polls_2008}
\title{Poll data for popular vote in 2008 presidential election}
\format{
An object of class \code{data.frame}.
}
\source{
\href{https://web.archive.org/web/20161108190914/http://www.pollster.com/08USPresGEMvO-2.html}{https://web.archive.org/web/20161108190914/http://www.pollster.com/08USPresGEMvO-2.html}
}
\usage{
polls_2008
}
\description{
Data from different pollsters for the popular vote between Obama and McCain 
in the 2008 presidential election.
}
\details{
\itemize{
\item day. Days until election day. Negative numbers are reported so that days can increase up to 0, which is election day.
\item margin. Average difference between Obama and McCain for that day.
}
}
\examples{
with(polls_2008, plot(day, margin))

}
\keyword{datasets}
