% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reported_heights.R
\docType{data}
\name{reported_heights}
\alias{reported_heights}
\title{Self-reported Heights}
\format{
An object of class \code{"data.frame"}.
}
\usage{
reported_heights
}
\description{
Students were asked to report their height (in inches) and sex in an online form. 
This table includes the results from four courses.
}
\details{
\itemize{
 \item time_stamp. Time and date of the entry.
 \item sex. Sex of the student.
 \item height. Height as reported by student by filling in a text free box. 
}
}
\examples{
head(reported_heights)
}
\keyword{datasets}
