% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trump_tweets.R
\docType{data}
\name{trump_tweets}
\alias{trump_tweets}
\alias{sentiment_counts}
\title{Trump Tweets from2009 to 2017}
\format{
An object of class \code{"data.frame"}.
}
\source{
The Trump Twitter Archive: \url{https://www.thetrumparchive.com/}
}
\usage{
trump_tweets
}
\description{
This dataset contains all tweets from Donald Trump's Twitter account from 2009 to 2017. Additionally, the results of a sentiment analysis, conducted on tweets from the campaign period (2015-06-17 to 2016-11-08), are included in \code{sentiment_counts}.
}
\details{
\itemize{
\item source. Device or service used to compose tweet.
\item id_str. Tweet ID.
\item text. Tweet.
\item created_at. Data and time tweet was tweeted.
\item retweet_count. How many times tweet had been retweeted at time dataset was created.
\item in_reply_to_user_id_str. If a reply, the user id of person being replied to.
\item favorite_count. Number of times tweet had been favored at time dataset was created.
\item is_retweet. A logical telling us if it is a retweet or not.
}
}
\examples{
head(trump_tweets)

}
\keyword{datasets}
