% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/us_contagious_disease.R
\docType{data}
\name{us_contagious_diseases}
\alias{us_contagious_diseases}
\title{Contagious disease data for US states}
\format{
An object of class \code{"data.frame"}.
}
\source{
\href{http://www.tycho.pitt.edu/}{Tycho Project}
}
\usage{
us_contagious_diseases
}
\description{
Yearly counts for Hepatitis A, Measles, Mumps, Pertussis, Polio,  Rubella, and Smallpox for US states. Original data courtesy of Tycho Project (http://www.tycho.pitt.edu/).
}
\details{
\itemize{
 \item disease. A factor containing disease names.
 \item state. A factor containing state names.
 \item year. 
 \item weeks_reporting. Number of weeks counts were reported that year.
 \item count. Total number of reported cases. 
 \item population. State population, interpolated for non-census years.
}
}
\examples{
head(us_contagious_diseases)
}
\references{
Willem G. van Panhuis, John Grefenstette, Su Yon Jung, Nian Shong Chok, Anne Cross, Heather Eng, Bruce Y Lee, Vladimir Zadorozhny, Shawn Brown, Derek Cummings, Donald S. Burke. Contagious Diseases in the United States from 1888 to the present. NEJM 2013; 369(22): 2152-2158.
}
\keyword{datasets}
